no process_raw_result_queue() anymore
There is a dedicated daemon
  benchmarkanything-storage-frontend-http
which does the queue processing in the
correct way, especially it takes care
to correctly handle db connections in
prefork situations.

The now droppped code here also did the
queue processing but is running in an
unknown fork context and did it not as
carefully as required.

Instead of rewriting the logic here we simply
require the daemon that already does it.